Upstream: Add convenience wrappers for common uio usage
The macOS uio struct is opaque and the API must be used, this
makes the smallest changes to the code for all platforms.

Most affected area is module/icp, which had a lot of code repetition.
It was tricky to make the for loop that skips to first non-empty
iov index look good, so it was moved into a macro. Technically, this
"skip code" is not required in ZFS, as all callers to crypto will not
create empty iov.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>