Extending FreeBSD UIO Struct
In FreeBSD the struct uio was just a typedef to uio_t. In order to
extend this struct, outside of the definition for the struct uio, the
struct uio has been embedded inside of a uio_t struct.

Also renamed all the uio_* interaces to be zfs_uio_* to make it clear
this is a ZFS interface.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>