Upstream: replace strchrnul()
A little torn as to removing this call, or adding it to SPL, but
since it is not used in a new special way, and would
in-efficiently write nil when not needed, the old style will do.
But flexible, say the word...

Apple has own fsync() due to some special considerations, which we are
hoping to remove with time.

Signed-off-by: Jorgen Lundman <lundman@lundman.net>