We need to ask IOMallocAligned for the enclosing POW2
vmem_create() arenas want at least natural alignment for
the spans they import, and will panic if they don't get it.

For sub-PAGESIZE calls to osif_malloc, align on PAGESIZE.
Otherwise align on the enclosing power of two.

Additional fiddling in spl-vmem.c to adjust the minimum
import sizes and quanta, and to re-enable a couple of
uncommented checks that would be used if we
vmem_xcreate().

wip

try to avoid misaligned panic

this works